backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
pillow
pymysql
python-multipart